requirements.txt
# FastAPI and ASGI server
fastapi==0.121.1
uvicorn[standard]==0.38.0

# Database
sqlalchemy[asyncio]==2.0.44
asyncpg==0.30.0
alembic==1.17.1
psycopg2-binary==2.9.11

# Admin
sqladmin==0.21.0

# Authentication
fastapi-users[sqlalchemy]==15.0.1
fastapi-users-db-sqlalchemy==7.0.0

# Caching
fastapi-cache2[redis]==0.2.2
redis==4.6.0

# Celery
celery[redis]==5.5.3

# Templates
jinja2==3.1.6

# Validation
pydantic==2.12.4
pydantic-settings==2.12.0

# HTTP Client
httpx==0.28.1

# Testing
pytest==9.0.0
pytest-asyncio==1.3.0
pytest-cov==7.0.0
aiosqlite==0.20.0

# CLI
typer==0.17.1
rich==14.2.0

# Development
python-dotenv==1.2.1
